LayoutTests/imported/w3c/web-platform-tests/trusted-types/eval-csp-tt-default-policy-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'require-trusted-types-for'.


FAIL eval of TrustedScript works. assert_equals: expected (number) 2 but got (object) object "1+1"
FAIL eval of string works. assert_equals: expected 5 but got 2
PASS eval of !TrustedScript and !string works.
PASS Function constructor of TrustedScript works.
FAIL Function constructor of string works. assert_equals: expected 5 but got 2
